% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_accu.R
\name{calculate_accu}
\alias{calculate_accu}
\title{Calculate accuracy}
\usage{
calculate_accu(beta, gammamat)
}
\arguments{
\item{beta}{The true vector of coefficients}

\item{gammamat}{The estimated gamma}
}
\value{
The estimated accuracy
}
\description{
Calculate accuracy
}
\examples{
calculate_accu(c(1,1,0), matrix(c(1,1,1,1,0,0), nrow = 2))
}
