% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alphag.R
\name{calculate_alphag}
\alias{calculate_alphag}
\title{Calculate alphag}
\usage{
calculate_alphag(
  rtri,
  X,
  Y,
  s,
  indvec,
  post,
  gam,
  gammaprime,
  sigma2,
  sigma2prime,
  indic,
  indv,
  g = 1
)
}
\arguments{
\item{rtri}{A right triangular matrix}

\item{X}{The design matrix}

\item{Y}{A vector of outcome}

\item{s}{A vector of variance of each column of X}

\item{indvec}{A vector indicates covariate}

\item{post}{The former posterior}

\item{gam}{A vector indicates whether has effect}

\item{gammaprime}{The proposal gammar}

\item{sigma2}{The variance}

\item{sigma2prime}{Th proposal variance}

\item{indic}{A number indicates which case}

\item{indv}{A number indicates which variate is changed}

\item{g}{A number indicates which balance function}
}
\value{
The value of alphag in log form
}
\description{
Calculate alphag
}
\examples{
rtri <- matrix(c(1,0,0,1,1,0,1,1,1), ncol = 3)
X <- matrix(1:9, ncol = 3)
calculate_alphag(rtri, X, c(1,1,1), c(1,1,1), c(1,2,3), 5, c(1,1,1), c(1,1,1), 1, 1.1, 1, 0, g = 1)
}
