% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{calculate_posterior}
\alias{calculate_posterior}
\title{Calculate posterior}
\usage{
calculate_posterior(rtri, X, Y, s, gam, sigma2)
}
\arguments{
\item{rtri}{A right triangular matrix}

\item{X}{Design matrix}

\item{Y}{A vector of outcome}

\item{s}{The vector of variance in each column of X}

\item{gam}{A vector indicates whether has effect}

\item{sigma2}{The variance}
}
\value{
The posterior in log form
}
\description{
Calculate posterior
}
\examples{
rtri <- matrix(c(1,0,0,1,1,0,1,1,1), ncol = 3)
X <- matrix(1:9, ncol = 3)
calculate_posterior(rtri, X, c(1,1,1), c(1,1,1), c(1,1,1), 1)
}
