% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_proposal.R
\name{draw_proposal}
\alias{draw_proposal}
\title{Draw proposal in each iteration of MCMC}
\usage{
draw_proposal(m, a, gam, sigma2, delta)
}
\arguments{
\item{m}{The number of trials}

\item{a}{a/m is the probability draw from the first case}

\item{gam}{A vector indicates whether has effect}

\item{sigma2}{The variance}

\item{delta}{The variance in first case}
}
\value{
A propsal
}
\description{
Draw proposal in each iteration of MCMC
}
\examples{
draw_proposal(5,1,c(1,0,1), 1, 1)
}
