% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(p, h, n, M)
}
\arguments{
\item{p}{The probability whether a covariate has effect}

\item{h}{The heritability}

\item{n}{Number of observations}

\item{M}{Total number of covariate}
}
\value{
Y, X, a vector of coefficients, random effect
}
\description{
Generate data
}
\examples{
generate_data(0.05, 0.5, 1000, 200)
}
