% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/given_rotation.R
\name{given_rotation}
\alias{given_rotation}
\title{Calculate values in given rotation matrix}
\usage{
given_rotation(a, b)
}
\arguments{
\item{a}{A number in the first entry}

\item{b}{A number in the second entry}
}
\value{
The values in given rotation matrix
}
\description{
Calculate values in given rotation matrix
}
\examples{
given_rotation(3, 4)
}
