% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_IT.R
\name{MT_IT}
\alias{MT_IT}
\title{The MT_IT methods}
\usage{
MT_IT(Y, X, m, a, delta, niter, g = 1)
}
\arguments{
\item{Y}{A n x 1 vector of outcome}

\item{X}{The n x M design matrix}

\item{m}{A value represents the number of trials in each iteration}

\item{a}{A number: a/m is the probability of purposing from first case}

\item{delta}{A value represents the variance in first case proposal}

\item{niter}{A value represents the number of iterations}

\item{g}{A value indicates the balance function}
}
\value{
A list with the elements
\item{gammamat}{A (niter+1) x M matrix contains gamma in each iteration}
\item{sigma2vec}{A (niter+1) x 1 vector contains sigma2 in each iteration}
\item{wmat}{A (niter+1) x 1 vector contains the weight in each iteration}
}
\description{
The MT_IT methods
}
\examples{
data <- generate_data(0.05, 0.4, 1000, 200)
  Y <- data$Y
  X <- data$X
  m <- 20
  a <- 2
  delta <- 0.5
  niter <- 1500
  res <- MT_IT(Y, X, m, a, delta, niter)
}
