% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatechol_add.R
\name{updatechol_add}
\alias{updatechol_add}
\title{Update Cholesky decomposition when adding a covariate}
\usage{
updatechol_add(rtri, sigma2, X, indvec, newv)
}
\arguments{
\item{rtri}{A p x p right triangular matrix}

\item{sigma2}{A number: the variance of coefficient}

\item{X}{The n x M design matrix}

\item{indvec}{A p x 1 vector indicates covariate}

\item{newv}{A number indicates new covariate}
}
\value{
The updated p x p right triangular matrix
}
\description{
Update Cholesky decomposition when adding a covariate
}
\examples{
rtri <- matrix(c(1,0,1,1), nrow = 2)
X <- matrix(c(1,0,1,1,1,1), nrow = 2)
updatechol_add(rtri, 1, X, c(1,2), 3)
}
