% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_accu.R
\name{calculate_accu}
\alias{calculate_accu}
\title{Calculate accuracy}
\usage{
calculate_accu(beta, gammamat)
}
\arguments{
\item{beta}{A M x 1 vector represents true vector of coefficients}

\item{gammamat}{A (niter+1) x M estimated gamma matrix}
}
\value{
A value represents estimated accuracy
}
\description{
Calculate accuracy
}
\examples{
calculate_accu(c(1,1,0), matrix(c(1,1,1,1,0,0), nrow = 2))
}
