% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alphag.R
\name{calculate_alphag}
\alias{calculate_alphag}
\title{Calculate alphag}
\usage{
calculate_alphag(
  rtri,
  X,
  Y,
  s,
  indvec,
  post,
  gam,
  gammaprime,
  sigma2,
  sigma2prime,
  indic,
  indv,
  g = 1
)
}
\arguments{
\item{rtri}{A p x p right triangular matrix}

\item{X}{A n x M Design matrix}

\item{Y}{A n x 1 vector of outcome}

\item{s}{A M x 1 vector for variance in each column of X}

\item{indvec}{A p x 1 vector indicates covariate that have effects}

\item{post}{A value represents the former posterior}

\item{gam}{A M x 1 vector indicates whether covariate has effect}

\item{gammaprime}{A M x 1 represents the proposal gamma}

\item{sigma2}{A value represents the variance}

\item{sigma2prime}{A value represents proposal variance}

\item{indic}{A number indicates which case}

\item{indv}{A number indicates which variate is changed}

\item{g}{A number indicates which balance function}
}
\value{
A list with the elements
\item{rtriprime}{The p x p right triangular matrix for the propsal}
\item{alpha}{A value of the function alphag in log form}
}
\description{
Calculate alphag
}
\examples{
rtri <- matrix(c(1,0,0,1,1,0,1,1,1), ncol = 3)
X <- matrix(1:9, ncol = 3)
calculate_alphag(rtri, X, c(1,1,1), c(1,1,1), c(1,2,3), 5, c(1,1,1), c(1,1,1), 1, 1.1, 1, 0, g = 1)
}
