% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{calculate_posterior}
\alias{calculate_posterior}
\title{Calculate posterior}
\usage{
calculate_posterior(rtri, X, Y, s, gam, sigma2)
}
\arguments{
\item{rtri}{A p x p right triangular matrix}

\item{X}{A n x M Design matrix}

\item{Y}{A n x 1 vector of outcome}

\item{s}{A M x 1 vector for variance in each column of X}

\item{gam}{A M x 1 vector indicates whether covariate has effect}

\item{sigma2}{A value represents the variance}
}
\value{
A value represents the posterior in log form
}
\description{
Calculate posterior
}
\examples{
rtri <- matrix(c(1,0,0,1,1,0,1,1,1), ncol = 3)
X <- matrix(1:9, ncol = 3)
calculate_posterior(rtri, X, c(1,1,1), c(1,1,1), c(1,1,1), 1)
}
