% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(p, h, n, M)
}
\arguments{
\item{p}{A number gives the probability whether a covariate has effect}

\item{h}{A number represents the heritability}

\item{n}{A scalar: number of observations}

\item{M}{A scalar: total number of covariates}
}
\value{
A list with the elements
\item{Y}{A n x 1 vector of outcomes}
\item{X}{A n x M matrix of the whole design matrix}
\item{beta}{A M x 1 vector for the true coefficients of each covariates}
\item{epsilon}{A n x 1 vector of random effects}
}
\description{
Generate data
}
\examples{
generate_data(0.05, 0.5, 1000, 200)
}
