% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/given_rotation.R
\name{given_rotation}
\alias{given_rotation}
\title{Calculate values in given rotation matrix}
\usage{
given_rotation(a, b)
}
\arguments{
\item{a}{A number(scalar) in the first entry}

\item{b}{A number(scalar) in the second entry}
}
\value{
A list with the elements
\item{c}{The value of c in th given rotation matrix.}
\item{s}{The value of s in th given rotation matrix.}
}
\description{
Calculate values in given rotation matrix
}
\examples{
given_rotation(3, 4)
}
