% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_IT.R
\name{MT_IT}
\alias{MT_IT}
\title{The MT_IT methods}
\usage{
MT_IT(Y, X, m, a, delta, niter, g = 1)
}
\arguments{
\item{Y}{A vector of outcome}

\item{X}{The design matrix}

\item{m}{The number of trials}

\item{a}{a/m the probability in first case proposal}

\item{delta}{The variance in first case proposal}

\item{niter}{Number of iterations}

\item{g}{The balance function}
}
\value{
The MCMC output
}
\description{
The MT_IT methods
}
\examples{
data <- generate_data(0.05, 0.4, 1000, 200)
  Y <- data$Y
  X <- data$X
  m <- 20
  a <- 2
  delta <- 0.5
  niter <- 1500
  res <- MT_IT(Y, X, m, a, delta, niter)
}
