% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatechol_remove.R
\name{updatechol_remove}
\alias{updatechol_remove}
\title{Update Cholesky Decomposition when removing a covariate}
\usage{
updatechol_remove(rtri, indvec, oldv)
}
\arguments{
\item{rtri}{A right triangular matrix}

\item{indvec}{A vector indicates the covariate}

\item{oldv}{A number indicates which covariate to be removed}
}
\value{
The new triangular matrix
}
\description{
Update Cholesky Decomposition when removing a covariate
}
\examples{
X <- matrix(rnorm(1000), ncol=20)
rtri <- chol(crossprod(X))
rtriprime <- updatechol_remove(rtri, 1:20, 5)
}
